refactor(delete-post): refactoring delete post button and access refs from provider
